Gh121 webui 2

+++ metasfresh-webui-api/src/main/java/de/metas/ui/web/pattribute/ASIRepository.java
@@ -151,7 +151,8 @@ public class ASIRepository
 		final DocumentId asiDocId = DocumentId.of(asiDocIdInt);
 		final int attributeSetId = getASIDocument(asiDocId)
 				.getEntityDescriptor()
-				.getDocumentTypeId();
+				.getDocumentTypeId()
+				.toInt();
 
 		return descriptorsFactory.getASIDescriptor(attributeSetId)
 				.getLayout();
@@ -243,7 +244,7 @@ public class ASIRepository
 	{
 		//
 		// Create M_AttributeSetInstance
-		final int attributeSetId = asiDoc.getEntityDescriptor().getDocumentTypeId();
+		final int attributeSetId = asiDoc.getEntityDescriptor().getDocumentTypeId().toInt();
 
 		final I_M_AttributeSetInstance asiRecord = InterfaceWrapperHelper.create(Env.getCtx(), I_M_AttributeSetInstance.class, ITrx.TRXNAME_ThreadInherited);
 		asiRecord.setM_AttributeSet_ID(attributeSetId);
